requirements.txt
pdfplumber
google-cloud-bigquery
torch
transformers
pytest
